correct gtol docs; ensure is int or None

+++ mystic/scipy_optimize.py
@@ -854,7 +854,7 @@ Args:
         convergence.
     ftol (float, default=1e-4): acceptable relative error in ``cost(xopt)``
         for convergence.
-    gtol (float, default=2): maximum iterations to run without improvement.
+    gtol (int, default=2): maximum iterations to run without improvement.
     maxiter (int, default=None): the maximum number of iterations to perform.
     maxfun (int, default=None): the maximum number of function evaluations.
     full_output (bool, default=False): True if fval and warnflag are desired.
